requirements.txt
 langchain
 langchain-anthropic
 python-dotenv
 wikipedia
 llm-math
 numexpr